+++ test/distributed/test_distributed.py
@@ -21,9 +21,14 @@ from torch.distributed.distributed_c10d import _get_default_group
 from torch._utils_internal import TEST_MASTER_ADDR as MASTER_ADDR
 from torch._utils_internal import TEST_MASTER_PORT as MASTER_PORT
 from torch.testing._internal.common_distributed import (
+    TEST_SKIPS,
     MultiProcessTestCase,
     simple_sparse_reduce_tests,
     skip_if_rocm,
+    skip_if_small_worldsize,
+    skip_if_no_cuda_distributed,
+    skip_if_lt_x_gpu,
+    skip_if_no_gpu,
 )
 
 try:
@@ -111,55 +116,6 @@ if not dist.is_available():
     sys.exit(0)
 
 
-SKIP_IF_NO_CUDA_EXIT_CODE = 75
-SKIP_IF_NO_GPU_EXIT_CODE = 76
-SKIP_IF_SMALL_WORLDSIZE_EXIT_CODE = 77
-SKIP_IF_BACKEND_UNAVAILABLE = 78
-SKIP_IF_ROCM_EXIT_CODE = 79
-
-
-def skip_if_no_cuda_distributed(func):
-    func.skip_if_no_cuda_distributed = True
-
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        if not torch.cuda.is_available():
-            sys.exit(SKIP_IF_NO_CUDA_EXIT_CODE)
-
-        return func(*args, **kwargs)
-
-    return wrapper
-
-
-def skip_if_no_gpu(func):
-    """ Nccl multigpu tests requires at least 2 GPUS. Skip if this is not met"""
-    func.skip_if_no_gpu = True
-
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        if not torch.cuda.is_available():
-            sys.exit(SKIP_IF_NO_CUDA_EXIT_CODE)
-        if torch.cuda.device_count() < int(os.environ["WORLD_SIZE"]):
-            sys.exit(SKIP_IF_NO_GPU_EXIT_CODE)
-
-        return func(*args, **kwargs)
-
-    return wrapper
-
-
-def skip_if_small_worldsize(func):
-    func.skip_if_small_worldsize = True
-
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        if (os.environ["BACKEND"] != "mpi") and int(os.environ["WORLD_SIZE"]) <= 2:
-            sys.exit(SKIP_IF_SMALL_WORLDSIZE_EXIT_CODE)
-
-        return func(*args, **kwargs)
-
-    return wrapper
-
-
 def skip_if_no_ninja(func):
 
     @wraps(func)
@@ -201,30 +157,6 @@ def require_world_size(world_size):
     if int(os.environ["WORLD_SIZE"]) < world_size:
         return unittest.skip("Test requires world size of %d" % world_size)
     return lambda func: func
-
-
-def require_num_gpus(n):
-    """
-    Require environment to have access to at least `n` GPUs.
-    Test is skipped otherwise.
-
-    Note: this check cannot run in the parent process, because calling
-    `torch.cuda.is_initialized()` will cause lazy initialization of a
-    CUDA runtime API context, and CUDA doesn't support forking.
-    """
-    def decorator(func):
-        func.skip_if_no_gpu = True
-
-        @wraps(func)
-        def wrapper(*args, **kwargs):
-            if not torch.cuda.is_available():
-                sys.exit(SKIP_IF_NO_CUDA_EXIT_CODE)
-            if torch.cuda.device_count() < n:
-                sys.exit(SKIP_IF_NO_GPU_EXIT_CODE)
-            return func(*args, **kwargs)
-        return wrapper
-
-    return decorator
 
 
 def apply_hack_for_nccl():
@@ -523,14 +455,14 @@ class _DistTestBase(object):
     @require_backend({"gloo", "nccl"})
     @require_backends_available({"gloo", "nccl"})
     @require_world_size(3)
-    @require_num_gpus(2)
+    @skip_if_lt_x_gpu(2)
     @skip_if_rocm
     def test_backend_group(self):
         self._test_group_override_backend(self._init_group_test)
 
     @require_backend({"gloo", "nccl"})
     @require_backends_available({"gloo", "nccl"})
-    @require_num_gpus(3)
+    @skip_if_lt_x_gpu(3)
     def test_backend_full_group(self):
         self._test_group_override_backend(self._init_full_group_test)
 
@@ -2312,8 +2244,8 @@ if BACKEND == "gloo" or BACKEND == "nccl":
                 )
             except RuntimeError as e:
                 if "recompile" in e.args[0]:
-                    sys.exit(SKIP_IF_BACKEND_UNAVAILABLE)
-                    # sys.exit(0)
+                    sys.exit(TEST_SKIPS["backend_unavailable"].exit_code)
+
                 raise
 
             # Execute barrier prior to running test to ensure that every process
